Refactor and rename files to follow the project conventions
