+++ registrants/src/UI/embc-registrant/src/app/core/components/dialog-components/self-serve-submissoin-dialog/self-serve-submission-dialog.component.html
@@ -1,0 +1,72 @@
+<div class="row" style="justify-content: flex-end">
+  <button class="close-image close-button-style" mat-icon-button aria-label="Close" mat-dialog-close>
+    <img src="/assets/images/close.svg" height="20" width="20" />
+    <img src="/assets/images/close_onhover.svg" height="20" width="20" />
+  </button>
+</div>
+<div style="display: flex; flex-direction: column; gap: var(--size-1); align-items: center">
+  <img src="../../../../assets/images/check-mark-icon.svg" style="width: 80px; height: 80px" alt="check_mark" />
+  <h1 mat-dialog-title>Support Submission Complete</h1>
+  <p class="card-heading">Your e-Transfer request is being processed.</p>
+</div>
+
+<mat-dialog-content>
+  <h2 style="font-size: var(--size-3); font-weight: bold">Next Steps</h2>
+  <p>1. <b>Amount:</b> $ {{ data?.totalAmount }}</p>
+  <p>
+    @switch (data?.notificationPreference) {
+      @case (ETransferNotificationPreference.Email) {
+        <div style="display: flex">
+          <div style="width: 240px">2. <b>e-Transfer Notification:</b></div>
+          <div style="width: 160px">
+            {{ data?.eTransferDetails?.eTransferEmail }}
+          </div>
+        </div>
+      }
+      @case (ETransferNotificationPreference.Mobile) {
+        <div style="display: flex">
+          <div style="width: 240px">2. <b>e-Transfer Notification:</b></div>
+          <div style="width: 160px">
+            {{ data?.eTransferDetails?.eTransferMobile }}
+          </div>
+        </div>
+        <div style="display: flex">
+          <div style="width: 240px">Contact Email Address</div>
+          <div style="width: 160px">
+            {{ data?.eTransferDetails?.contactEmail }}
+          </div>
+        </div>
+      }
+      @case (ETransferNotificationPreference.EmailAndMobile) {
+        <div style="display: flex">
+          <div style="width: 240px">2. <b>e-Transfer Notification:</b></div>
+          <div style="width: 160px">
+            {{ data?.eTransferDetails?.eTransferMobile }}
+          </div>
+        </div>
+        <div style="display: flex">
+          <div style="width: 240px">&nbsp;</div>
+          <div style="width: 160px">
+            {{ data?.eTransferDetails?.eTransferEmail }}
+          </div>
+        </div>
+      }
+    }
+  </p>
+  <p>
+    3. <b>Support Contact:</b> While most e-Transfers are processed immediately, there are instances where processing
+    delays can occur. If your e-Transfer is not received within 24 hours, please contact <br />
+    <a>1-800-585-9559</a>.
+  </p>
+  <p>
+    4. <b>No Further Action Required:</b> You DO NOT need to meet with a responder at a reception centre. For
+    information about evacuation orders and alerts, please visit your Local Emergency Programs website or
+    <br />
+    <a href="https://EmergencyInfoBC.ca" target="blank">EmergencyInfoBC.ca</a>.
+  </p>
+  <p>5. <b>Email Confirmation:</b> A confirmation email will be sent to the email on file for your records.</p>
+</mat-dialog-content>
+
+<mat-dialog-actions style="justify-content: center">
+  <button class="button-p" style="max-width: 280px" cdkFocusInitial mat-button mat-dialog-close>Close</button>
+</mat-dialog-actions>
