+++ registrants/src/UI/embc-registrant/src/app/core/components/dialog-components/self-serve-submissoin-dialog/self-serve-submission-dialog.component.scss
@@ -1,0 +1,8 @@
+:host {
+  display: block;
+  padding: var(--size-3);
+}
+
+:host p {
+  margin-bottom: var(--size-2);
+}
