--- registrants/src/UI/embc-registrant/src/app/feature-components/eligible-self-serve/eligible-self-serve.routes.ts
@@ -1,9 +1,0 @@
-import { EligibleSelfServeConfirmComponent } from './eligible-self-serve-confirm.component';
-import { EligibleSelfServeSupportFormComponent } from './eligible-self-serve-support-form.component';
-import { Routes } from '@angular/router';
-
-export const routes: Routes = [
-  { path: '', redirectTo: 'confirm', pathMatch: 'full' },
-  { path: 'confirm', component: EligibleSelfServeConfirmComponent },
-  { path: 'support-form', component: EligibleSelfServeSupportFormComponent }
-];
