+++ registrants/src/UI/embc-registrant/src/app/feature-components/self-serve-support/self-serve-support-confirm/self-serve-support-confirm.component.html
@@ -7,7 +7,7 @@
       </h1>
 
       <h2 class="highlightText">
-        Before proceeding, please <b></b>confirm that you meet the following <b>requirements</b>.
+        Before proceeding, please <b>confirm</b> that you meet the following <b>requirements</b>.
       </h2>
 
       <img src="assets/images/valid_bank_account.svg" width="200" height="200" />
