+++ registrants/src/UI/embc-registrant/src/app/feature-components/self-serve-support/self-serve-support-confirm/self-serve-support-confirm.component.ts
@@ -2,23 +2,23 @@ import { CUSTOM_ELEMENTS_SCHEMA, Component } from '@angular/core';
 import { MatCardModule } from '@angular/material/card';
 import { MatDialog } from '@angular/material/dialog';
 import { ActivatedRoute, Router } from '@angular/router';
-import { EligibleSelfServeOptOutDialogComponent } from './eligible-self-serve-opt-out-dialog.component';
+import { SelfServeSupportOptOutDialogComponent } from '../self-serve-support-opt-out-dialog/self-serve-support-opt-out-dialog.component';
 import { MatButtonModule } from '@angular/material/button';
 import { DialogContent } from 'src/app/core/model/dialog-content.model';
 import { InformationDialogComponent } from 'src/app/core/components/dialog-components/information-dialog/information-dialog.component';
 import { DialogComponent } from 'src/app/core/components/dialog/dialog.component';
-import * as globalConst from '../../core/services/globalConstants';
+import * as globalConst from '../../../core/services/globalConstants';
 import { SupportsService } from 'src/app/core/api/services';
-import { NeedsAssessmentService } from '../needs-assessment/needs-assessment.service';
+import { NeedsAssessmentService } from '../../needs-assessment/needs-assessment.service';
 
 @Component({
   standalone: true,
-  selector: 'app-eligible-self-serve-confirm',
-  templateUrl: './eligible-self-serve-confirm.component.html',
+  selector: 'app-self-serve-confirm',
+  templateUrl: './self-serve-support-confirm.component.html',
   imports: [MatCardModule, MatButtonModule],
   schemas: [CUSTOM_ELEMENTS_SCHEMA]
 })
-export class EligibleSelfServeConfirmComponent {
+export class SelfServeSupportConfirmComponent {
   essFileId = this.needsAssessmentService.getVerifiedEvacuationFileNo();
 
   constructor(
@@ -31,7 +31,7 @@ export class EligibleSelfServeConfirmComponent {
 
   optOutSelfServe() {
     this.dialog
-      .open(EligibleSelfServeOptOutDialogComponent, {
+      .open(SelfServeSupportOptOutDialogComponent, {
         width: '700px'
       })
       .afterClosed()
