+++ registrants/src/UI/embc-registrant/src/app/feature-components/self-serve-support/self-serve-support-form.component.scss
@@ -1,0 +1,22 @@
+mat-card {
+  margin-bottom: var(--size-3);
+}
+
+.mat-mdc-card-content,
+.mat-mdc-card-content:first-child,
+.mat-mdc-card-content:last-child {
+  display: flex;
+  flex-wrap: wrap;
+  gap: var(--size-3);
+  justify-content: space-between;
+  padding: var(--size-3);
+}
+
+:host .mat-mdc-checkbox ::ng-deep label {
+  font-size: var(--size-2);
+  padding: 4px 0;
+}
+
+:host .mat-mdc-checkbox ::ng-deep .mdc-form-field {
+  align-items: flex-start;
+}
