+++ registrants/src/UI/embc-registrant/src/app/feature-components/self-serve-support/self-serve-support-total-amount-zero-dialog/self-serve-support-total-amount-zero.component.ts
@@ -2,44 +2,14 @@ import { Component } from '@angular/core';
 import { MatButtonModule } from '@angular/material/button';
 import { MatDialogModule, MatDialogRef } from '@angular/material/dialog';
 import { SupportsService } from 'src/app/core/api/services';
-import { NeedsAssessmentService } from '../needs-assessment/needs-assessment.service';
+import { NeedsAssessmentService } from '../../needs-assessment/needs-assessment.service';
 import { Router } from '@angular/router';
 
 @Component({
   standalone: true,
   selector: 'app-eligible-self-serve-opt-out-dialog',
   imports: [MatDialogModule, MatButtonModule],
-  template: `
-    <div class="row">
-      <div class="col-10">
-        <h1 mat-dialog-title>Support Update Required</h1>
-      </div>
-      <div class="col-2" style="text-align: right;">
-        <button class="close-image close-button-style" mat-icon-button aria-label="Close" mat-dialog-close>
-          <img src="/assets/images/close.svg" height="20" width="20" />
-          <img src="/assets/images/close_onhover.svg" height="20" width="20" />
-        </button>
-      </div>
-    </div>
-
-    <mat-dialog-content>
-      <p>
-        <b>You currently have no supports selected.</b> Please take a moment to update your support details to select
-        the necessary assistance for you and your household while you are evacuated.
-      </p>
-      <p>
-        If you no longer require support or would prefer to proceed with referrals, you can
-
-        <a class="highlightText" style="cursor: pointer" (click)="optOut()">
-          click here to opt out and receive supports via referral.
-        </a>
-      </p>
-    </mat-dialog-content>
-
-    <mat-dialog-actions>
-      <button class="button-p" mat-button mat-dialog-close>Edit Support Details</button>
-    </mat-dialog-actions>
-  `,
+  template: './self-serve-support-total-amount-zero.component.html',
   styles: [
     `
       :host {
