+++ registrants/src/UI/embc-registrant/src/app/feature-components/self-serve-support/self-serve-support.routes.ts
@@ -1,0 +1,9 @@
+import { SelfServeSupportConfirmComponent } from './self-serve-support-confirm/self-serve-support-confirm.component';
+import { SelfServeSupportFormComponent } from './self-serve-support-form.component';
+import { Routes } from '@angular/router';
+
+export const routes: Routes = [
+  { path: '', redirectTo: 'confirm', pathMatch: 'full' },
+  { path: 'confirm', component: SelfServeSupportConfirmComponent },
+  { path: 'support-form', component: SelfServeSupportFormComponent }
+];
